Source Truth/requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.2.0
numpy>=1.26.0
gspread>=6.1.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
openai>=1.30.0
langchain-openai>=0.2.0
